+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/Cohost/DidOpenHandler.cs
@@ -18,34 +18,20 @@ namespace Microsoft.VisualStudio.LanguageServerClient.Razor.Cohost;
 internal class DidOpenHandler(
     ProjectSnapshotManagerDispatcher projectSnapshotManagerDispatcher,
     RazorProjectService razorProjectService,
-    ISnapshotResolver snapshotResolver,
     OpenDocumentGenerator openDocumentGenerator) : IRazorCohostDidOpenHandler
 {
     private readonly ProjectSnapshotManagerDispatcher _projectSnapshotManagerDispatcher = projectSnapshotManagerDispatcher;
     private readonly RazorProjectService _razorProjectService = razorProjectService;
-    private readonly ISnapshotResolver _snapshotResolver = snapshotResolver;
     private readonly OpenDocumentGenerator _openDocumentGenerator = openDocumentGenerator;
 
     public async Task HandleAsync(Uri uri, int version, SourceText sourceText, CancellationToken cancellationToken)
     {
-        await await _projectSnapshotManagerDispatcher.RunOnDispatcherThreadAsync(async () =>
+        await await _projectSnapshotManagerDispatcher.RunOnDispatcherThreadAsync(() =>
         {
             var textDocumentPath = FilePathNormalizer.Normalize(uri.GetAbsoluteOrUNCPath());
             _razorProjectService.OpenDocument(textDocumentPath, sourceText, version);
 
-            if (!_snapshotResolver.TryResolveAllProjects(textDocumentPath, out var projectSnapshots))
-            {
-                projectSnapshots = [_snapshotResolver.GetMiscellaneousProject()];
-            }
-
-            foreach (var project in projectSnapshots)
-            {
-                var document = project.GetDocument(textDocumentPath);
-                if (document is not null)
-                {
-                    await _openDocumentGenerator.DocumentOpenedOrChangedAsync(document, version, cancellationToken);
-                }
-            }
+            return _openDocumentGenerator.DocumentOpenedOrChangedAsync(textDocumentPath, version, cancellationToken);
         }, cancellationToken).ConfigureAwait(false);
     }
 }
